Read ventoy parameter from acpi firstly to avoid kernel-lockdown protect. Ventoy issue 714

+++ build.sh
@@ -8,8 +8,16 @@ rm -f vtoydump*
 gcc -Wall -std=gnu99 -DHAVE_CONFIG_H  -O2 -D_FILE_OFFSET_BITS=64 ./src/vtoydump_linux.c ./src/libexfat/*.c -I ./src -I ./src/libexfat -o vtoydump64
 gcc -Wall -std=gnu99 -DHAVE_CONFIG_H  -O2 -D_FILE_OFFSET_BITS=64 -m32 ./src/vtoydump_linux.c ./src/libexfat/*.c -I ./src -I ./src/libexfat -o vtoydump32
 
-if [ -e vtoydump64 ] && [ -e vtoydump32 ]; then
+aarch64-linux-gnu-gcc -Wall -std=gnu99 -DHAVE_CONFIG_H  -O2 -D_FILE_OFFSET_BITS=64 ./src/vtoydump_linux.c ./src/libexfat/*.c -I ./src -I ./src/libexfat -o vtoydumpaa64
+
+if [ -e vtoydump64 ] && [ -e vtoydump32 ] && [ -e vtoydumpaa64 ]; then
     rm -f ../Vtoyboot/vtoyboot/tools/vtoydump*
+    
+    strip vtoydump64
+    strip vtoydump32
+    aarch64-linux-gnu-strip vtoydumpaa64
+    
+    cp -a vtoydumpaa64 ../Vtoyboot/vtoyboot/tools/
     cp -a vtoydump64 ../Vtoyboot/vtoyboot/tools/
     cp -a vtoydump32 ../Vtoyboot/vtoyboot/tools/
